splited features
