+++ packages/espressocash_app/lib/features/settings/widgets/edit_profile_section.dart
@@ -1,13 +1,13 @@
-part of 'profile_section.dart';
+part of 'settings_section.dart';
 
 class EditProfileSection extends StatelessWidget {
   const EditProfileSection({super.key});
 
   @override
-  Widget build(BuildContext context) => ProfileSection(
+  Widget build(BuildContext context) => SettingsSection(
         padding: const EdgeInsets.fromLTRB(8, 16, 2, 16),
         actions: [
-          ProfileButton(
+          SettingsButton(
             label: context.l10n.editProfile,
             onPressed: () => ManageProfileScreen.push(context),
           ),
