+++ packages/espressocash_app/lib/features/settings/widgets/security_section.dart
@@ -7,31 +7,31 @@ import '../../app_lock/widgets/app_lock_menu_item.dart';
 import '../../legal/privacy_screen.dart';
 import '../../legal/terms_screen.dart';
 import '../../view_phrase/view_phrase.dart';
-import 'profile_button.dart';
-import 'profile_section.dart';
+import 'settings_button.dart';
+import 'settings_section.dart';
 
 class SecuritySection extends StatelessWidget {
   const SecuritySection({super.key});
 
   @override
-  Widget build(BuildContext context) => ProfileSection(
+  Widget build(BuildContext context) => SettingsSection(
         title: context.l10n.securitySectionTitle,
         padding: const EdgeInsets.fromLTRB(8, 16, 2, 16),
         actions: [
           const AppLockMenuItem(),
           ValueListenableBuilder(
             valueListenable: sl<AccountService>(),
-            builder: (context, value, child) => ProfileButton(
+            builder: (context, value, child) => SettingsButton(
               label: context.l10n.viewRecoveryPhrase,
               description: context.l10n.viewRecoveryPhraseDescription,
               onPressed: () => context.launchViewRecoveryPhraseFlow(),
             ),
           ),
-          ProfileButton(
+          SettingsButton(
             label: context.l10n.termsOfUse,
             onPressed: () => TermsScreen.push(context),
           ),
-          ProfileButton(
+          SettingsButton(
             label: context.l10n.privacyPolicy,
             onPressed: () => PrivacyScreen.push(context),
           ),
