+++ packages/espressocash_app/lib/features/settings/widgets/settings_section.dart
@@ -13,16 +13,16 @@ import '../../../ui/loader.dart';
 import '../../../ui/snackbar.dart';
 import '../../../ui/web_view_screen.dart';
 import '../../accounts/services/account_service.dart';
-import '../screens/manage_profile_screen.dart';
-import 'profile_button.dart';
+import '../../profile/screens/manage_profile_screen.dart';
+import 'settings_button.dart';
 
 part 'danger_section.dart';
 part 'edit_profile_section.dart';
 part 'share_section.dart';
 part 'version_section.dart';
 
-class ProfileSection extends StatelessWidget {
-  const ProfileSection({
+class SettingsSection extends StatelessWidget {
+  const SettingsSection({
     super.key,
     this.title,
     required this.actions,
